==> notebooks/1_automobile_preprocessing.ipynb <==
# Предобработка датасета automobile

Загрузка библиотек

import pandas as pd

Загрузка датасета

data = pd.read_csv(r'../data/automobile.csv', sep = ',')
data

data.info()
cols = data.columns
for col in cols:
    print(f"Признак: {col}, значения: {set(data[col].tolist())}")

##### Перекодирование

data = data.replace('?', '')
data = data.replace('two', 2)
data = data.replace('three', 3)
data = data.replace('four', 4)
data = data.replace('five', 5)
data = data.replace('six', 6)
data = data.replace('eight', 8)
data = data.replace('twelve', 12)

data.to_csv(r'../data/automobile_preprocessed.csv', index_label=False, index=False)
data = pd.read_csv(r'../data/automobile_preprocessed.csv', sep = ',')

data = data.dropna()

data.info()
data

cols = data.columns
for col in cols:
    print(f"Признак: {col}, значения: {set(data[col].tolist())}")

data = pd.get_dummies(data)

data.info()
data

##### Сохранение изменённого датасета

data.to_csv(r'../data/automobile_preprocessed.csv', index_label=False, index=False)
