==> notebooks/1_car-evaluation_preprocessing.ipynb <==
# Предобработка датасета car-evaluation

Загрузка библиотек

import pandas as pd

Загрузка датасета

data = pd.read_csv(r'../data/car-evaluation.csv', sep = ',')
data

data.info()
cols = data.columns
for col in cols:
    print(f"Признак: {col}, значения: {set(data[col].tolist())}")

##### Перекодирование

data.at[data['buying'] == 'low','buying'] = 1
data.at[data['buying'] == 'med','buying'] = 2
data.at[data['buying'] == 'high','buying'] = 3
data.at[data['buying'] == 'vhigh','buying'] = 4

data.at[data['maint'] == 'low','maint'] = 1
data.at[data['maint'] == 'med','maint'] = 2
data.at[data['maint'] == 'high','maint'] = 3
data.at[data['maint'] == 'vhigh','maint'] = 4

data.at[data['doors'] == '5more','doors'] = 5

data.at[data['persons'] == 'more','persons'] = 5

data.at[data['lug_boot'] == 'small','lug_boot'] = 1
data.at[data['lug_boot'] == 'med','lug_boot'] = 2
data.at[data['lug_boot'] == 'big','lug_boot'] = 3

data.at[data['safety'] == 'low','safety'] = 1
data.at[data['safety'] == 'med','safety'] = 2
data.at[data['safety'] == 'high','safety'] = 3

data.at[data['class'] == 'unacc','class'] = 1
data.at[data['class'] == 'acc','class'] = 2
data.at[data['class'] == 'good','class'] = 3
data.at[data['class'] == 'vgood','class'] = 4

data

data.info()
cols = data.columns
for col in cols:
    print(f"Признак: {col}, значения: {set(data[col].tolist())}")

##### Сохранение изменённого датасета

data.to_csv(r'../data/car-evaluation_preprocessed.csv', index_label=False, index=False)
