==> notebooks/3_automobile_regression.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv(r'../data/automobile_preprocessed.csv')

data.info()
data

Выделим вектор исследуемого признака и матрицу объект-признак

y = data['symboling']
y

cols = data.columns[1:60]
X = data[cols]
X

Разбиваем выборку

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error

# Линейная регрессия

from sklearn.linear_model import LinearRegression

lr = LinearRegression()
lr.fit(X_train, y_train)
y_pred_lr = lr.predict(X_test)

Оценим качество модели

print(r2_score(y_test, y_pred_lr))
print(mean_squared_error(y_test, y_pred_lr))
print(mean_squared_error(y_test, y_pred_lr, squared=False))
print(mean_absolute_error(y_test, y_pred_lr))

# Регрессия дерева решений

from sklearn.tree import DecisionTreeRegressor

dtr = DecisionTreeRegressor()
dtr.fit(X_train, y_train)
y_pred_dtr = dtr.predict(X_test)

Оценим качество модели

print(r2_score(y_test, y_pred_dtr))
print(mean_squared_error(y_test, y_pred_dtr))
print(mean_squared_error(y_test, y_pred_dtr, squared=False))
print(mean_absolute_error(y_test, y_pred_dtr))

# LASSO

from sklearn.linear_model import Lasso

lasso = Lasso()
lasso.fit(X_train, y_train)
y_pred_lasso = lasso.predict(X_test)

Оценим качество модели

print(r2_score(y_test, y_pred_lasso))
print(mean_squared_error(y_test, y_pred_lasso))
print(mean_squared_error(y_test, y_pred_lasso, squared=False))
print(mean_absolute_error(y_test, y_pred_lasso))

#  Гребневая регрессия

from sklearn.linear_model import Ridge

ridge = Ridge()
ridge.fit(X_train,y_train)
y_pred_ridge = ridge.predict(X_test)

Оценим качество модели

print(r2_score(y_test, y_pred_ridge))
print(mean_squared_error(y_test, y_pred_ridge))
print(mean_squared_error(y_test, y_pred_ridge, squared=False))
print(mean_absolute_error(y_test, y_pred_ridge))

#  Elastic Net регрессия

from sklearn.linear_model import ElasticNet

en = ElasticNet()
en.fit(X_train, y_train)
y_pred_en = en.predict(X_test)

Оценим качество модели

print(r2_score(y_test, y_pred_en))
print(mean_squared_error(y_test, y_pred_en))
print(mean_squared_error(y_test, y_pred_en, squared=False))
print(mean_absolute_error(y_test, y_pred_en))

# Вывод
Лучший результат показала регрессия дерева решений
