==> notebooks/4_ensemble-classification.ipynb <==
Импорт библиотек и загрузка датасета

import pandas as pd
import numpy as np
import warnings
from sklearn.metrics import confusion_matrix, classification_report, mean_squared_error, mean_absolute_error
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import BaggingClassifier, GradientBoostingClassifier, StackingClassifier

warnings.filterwarnings("ignore")

data = pd.read_csv(r'../data/car-evaluation_preprocessed.csv', sep=',')

data

X = data.iloc[:, :-1].values
y = data.iloc[:, 6].values


X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

def print_classification_model_metrics(estimator, y_test, y_pred):
    print(confusion_matrix(y_test, y_pred))
    print(classification_report(y_test, y_pred))
    print(estimator.score(X_test, y_test))

# DTC

%%time
parameters_dtc = {'max_depth': np.arange(1,20,1)}
dtc = DecisionTreeClassifier()
dtc_base = GridSearchCV(dtc, parameters_dtc).fit(X_train, y_train)
dtc_base.best_params_

print_classification_model_metrics(dtc_base, y_test, dtc_base.predict(X_test))

ensemble params

parameters_ensemble = {'n_estimators': np.arange(0,100,5),
                       'max_features': np.arange(1,24,3)}

# Бэггинг

%%time
bag = BaggingClassifier(base_estimator=DecisionTreeClassifier(max_depth=dtc_base.best_params_['max_depth']))
model = GridSearchCV(bag, parameters_ensemble).fit(X_train, y_train)
model.best_params_

print_classification_model_metrics(model, y_test, model.predict(X_test))

# Бустинг

%%time
gbc = GradientBoostingClassifier()
model = GridSearchCV(gbc, parameters_ensemble).fit(X_train, y_train)
model.best_params_

print_classification_model_metrics(model, y_test, model.predict(X_test))

# Стекинг

%%time
model = StackingClassifier(estimators=[('bag',bag), ('gbc',gbc)],
                           final_estimator=dtc_base).fit(X_train, y_train)

print_classification_model_metrics(model, y_test, model.predict(X_test))
