==> notebooks/4_ensemble-regression.ipynb <==
Импорт библиотек и загрузка датасета

import pandas as pd
import numpy as np
import warnings
from sklearn.metrics import confusion_matrix, classification_report, mean_squared_error, mean_absolute_error
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import BaggingRegressor, GradientBoostingRegressor, StackingRegressor

warnings.filterwarnings("ignore")

data = pd.read_csv(r'../data/automobile_preprocessed.csv')

data

cols = data.columns[1:60]
X = data[cols]
X

y = data['symboling']
y

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.40)

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

def print_regression_model_metrics(estimator, y_test, y_pred):
    print(f"Коэффициент детерминации: {estimator.score(X,y)}")
    print(f'MSE: {mean_squared_error(y_test, y_pred)}')
    print(f'RMSE: {mean_squared_error(y_test, y_pred, squared=False)}')
    print(f'MAE: {mean_absolute_error(y_test, y_pred)}')

# DTR

%%time
parameters_dtr = {'max_depth': np.arange(5,20,1)}
dtr = DecisionTreeRegressor()
dtr_base = GridSearchCV(dtr, parameters_dtr).fit(X_train, y_train)
dtr_base.best_params_
print_regression_model_metrics(dtr_base, y_test, dtr_base.predict(X_test))

ensemble params

parameters_ensemble = {'n_estimators': np.arange(0,100,5),
                       'max_features': np.arange(1,24,3)}

# Бэггинг

%%time
bag = BaggingRegressor(base_estimator=DecisionTreeRegressor(max_depth=dtr_base.best_params_['max_depth']))
model = GridSearchCV(bag, parameters_ensemble).fit(X_train, y_train)
model.best_params_

print_regression_model_metrics(model, y_test, model.predict(X_test))

# Бустинг

%%time
gbc = GradientBoostingRegressor()
model = GridSearchCV(gbc, parameters_ensemble).fit(X_train, y_train)
model.best_params_

print_regression_model_metrics(model, y_test, model.predict(X_test))

# Стекинг

%%time
model = StackingRegressor(estimators=[('bag',bag), ('gbc',gbc)],
                           final_estimator=dtr_base).fit(X_train, y_train)

print_regression_model_metrics(model, y_test, model.predict(X_test))
